Adds module for utils and data creation. Refactors getRandomInteger and getRandomArray utils
